+++ tests/Catalog/CatalogServiceTest.php
@@ -342,6 +342,14 @@ final class CatalogServiceTest extends ApiTestCase
                         'variantName' => '13',
                     ],
                 ],
+                'company' => [
+                    'id' => 6,
+                    'name' => 'Test company',
+                    'slug' => 'test-company-2',
+                    'isProfessional' => true,
+                    'image' => null,
+                    'averageRating' => null,
+                ],
             ]),
             new Declination([
                 'id' => '2_7_2',
@@ -366,6 +374,14 @@ final class CatalogServiceTest extends ApiTestCase
                         'variantId' => 2,
                         'variantName' => '15',
                     ],
+                ],
+                'company' => [
+                    'id' => 6,
+                    'name' => 'Test company',
+                    'slug' => 'test-company-2',
+                    'isProfessional' => true,
+                    'image' => null,
+                    'averageRating' => null,
                 ],
             ]),
             new Declination([
@@ -392,6 +408,14 @@ final class CatalogServiceTest extends ApiTestCase
                         'variantName' => '17',
                     ],
                 ],
+                'company' => [
+                    'id' => 6,
+                    'name' => 'Test company',
+                    'slug' => 'test-company-2',
+                    'isProfessional' => true,
+                    'image' => null,
+                    'averageRating' => null,
+                ],
             ]),
             new Declination([
                 'id' => '2_7_4',
@@ -416,6 +440,14 @@ final class CatalogServiceTest extends ApiTestCase
                         'variantId' => 4,
                         'variantName' => '21',
                     ],
+                ],
+                'company' => [
+                    'id' => 6,
+                    'name' => 'Test company',
+                    'slug' => 'test-company-2',
+                    'isProfessional' => true,
+                    'image' => null,
+                    'averageRating' => null,
                 ],
             ]),
         ];
@@ -484,6 +516,14 @@ final class CatalogServiceTest extends ApiTestCase
                         'variantName' => 'white',
                     ],
                 ],
+                'company' => [
+                    'id' => 7,
+                    'name' => 'Test company',
+                    'slug' => 'test-company-3',
+                    'isProfessional' => true,
+                    'image' => null,
+                    'averageRating' => null,
+                ],
             ]),
             new Declination([
                 'id' => '3_9_5_10_9',
@@ -514,6 +554,14 @@ final class CatalogServiceTest extends ApiTestCase
                         'variantId' => 9,
                         'variantName' => 'wireless',
                     ],
+                ],
+                'company' => [
+                    'id' => 7,
+                    'name' => 'Test company',
+                    'slug' => 'test-company-3',
+                    'isProfessional' => true,
+                    'image' => null,
+                    'averageRating' => null,
                 ],
             ]),
             new Declination([
@@ -546,6 +594,14 @@ final class CatalogServiceTest extends ApiTestCase
                         'variantName' => 'wired',
                     ],
                 ],
+                'company' => [
+                    'id' => 7,
+                    'name' => 'Test company',
+                    'slug' => 'test-company-3',
+                    'isProfessional' => true,
+                    'image' => null,
+                    'averageRating' => null,
+                ],
             ]),
             new Declination([
                 'id' => '3_9_6',
@@ -570,6 +626,14 @@ final class CatalogServiceTest extends ApiTestCase
                         'variantId' => 6,
                         'variantName' => 'black',
                     ],
+                ],
+                'company' => [
+                    'id' => 7,
+                    'name' => 'Test company',
+                    'slug' => 'test-company-3',
+                    'isProfessional' => true,
+                    'image' => null,
+                    'averageRating' => null,
                 ],
             ]),
             new Declination([
@@ -602,6 +666,14 @@ final class CatalogServiceTest extends ApiTestCase
                         'variantName' => 'wireless',
                     ],
                 ],
+                'company' => [
+                    'id' => 7,
+                    'name' => 'Test company',
+                    'slug' => 'test-company-3',
+                    'isProfessional' => true,
+                    'image' => null,
+                    'averageRating' => null,
+                ],
             ]),
             new Declination([
                 'id' => '3_9_6_10_10',
@@ -633,6 +705,14 @@ final class CatalogServiceTest extends ApiTestCase
                         'variantName' => 'wired',
                     ],
                 ],
+                'company' => [
+                    'id' => 7,
+                    'name' => 'Test company',
+                    'slug' => 'test-company-3',
+                    'isProfessional' => true,
+                    'image' => null,
+                    'averageRating' => null,
+                ],
             ]),
             new Declination([
                 'id' => '3_9_7',
@@ -657,6 +737,14 @@ final class CatalogServiceTest extends ApiTestCase
                         'variantId' => 7,
                         'variantName' => 'blue',
                     ],
+                ],
+                'company' => [
+                    'id' => 7,
+                    'name' => 'Test company',
+                    'slug' => 'test-company-3',
+                    'isProfessional' => true,
+                    'image' => null,
+                    'averageRating' => null,
                 ],
             ]),
             new Declination([
@@ -689,6 +777,14 @@ final class CatalogServiceTest extends ApiTestCase
                         'variantName' => 'wireless',
                     ],
                 ],
+                'company' => [
+                    'id' => 7,
+                    'name' => 'Test company',
+                    'slug' => 'test-company-3',
+                    'isProfessional' => true,
+                    'image' => null,
+                    'averageRating' => null,
+                ],
             ]),
             new Declination([
                 'id' => '3_9_7_10_10',
@@ -720,6 +816,14 @@ final class CatalogServiceTest extends ApiTestCase
                         'variantName' => 'wired',
                     ],
                 ],
+                'company' => [
+                    'id' => 7,
+                    'name' => 'Test company',
+                    'slug' => 'test-company-3',
+                    'isProfessional' => true,
+                    'image' => null,
+                    'averageRating' => null,
+                ],
             ]),
             new Declination([
                 'id' => '3_9_8',
@@ -744,6 +848,14 @@ final class CatalogServiceTest extends ApiTestCase
                         'variantId' => 8,
                         'variantName' => 'red',
                     ],
+                ],
+                'company' => [
+                    'id' => 7,
+                    'name' => 'Test company',
+                    'slug' => 'test-company-3',
+                    'isProfessional' => true,
+                    'image' => null,
+                    'averageRating' => null,
                 ],
             ]),
             new Declination([
@@ -776,6 +888,14 @@ final class CatalogServiceTest extends ApiTestCase
                         'variantName' => 'wireless',
                     ],
                 ],
+                'company' => [
+                    'id' => 7,
+                    'name' => 'Test company',
+                    'slug' => 'test-company-3',
+                    'isProfessional' => true,
+                    'image' => null,
+                    'averageRating' => null,
+                ],
             ]),
             new Declination([
                 'id' => '3_9_8_10_10',
@@ -807,6 +927,14 @@ final class CatalogServiceTest extends ApiTestCase
                         'variantName' => 'wired',
                     ],
                 ],
+                'company' => [
+                    'id' => 7,
+                    'name' => 'Test company',
+                    'slug' => 'test-company-3',
+                    'isProfessional' => true,
+                    'image' => null,
+                    'averageRating' => null,
+                ],
             ]),
             new Declination([
                 'id' => '3_10_9',
@@ -832,6 +960,14 @@ final class CatalogServiceTest extends ApiTestCase
                         'variantName' => 'wireless',
                     ],
                 ],
+                'company' => [
+                    'id' => 7,
+                    'name' => 'Test company',
+                    'slug' => 'test-company-3',
+                    'isProfessional' => true,
+                    'image' => null,
+                    'averageRating' => null,
+                ],
             ]),
             new Declination([
                 'id' => '3_10_10',
@@ -856,6 +992,14 @@ final class CatalogServiceTest extends ApiTestCase
                         'variantId' => 10,
                         'variantName' => 'wired',
                     ],
+                ],
+                'company' => [
+                    'id' => 7,
+                    'name' => 'Test company',
+                    'slug' => 'test-company-3',
+                    'isProfessional' => true,
+                    'image' => null,
+                    'averageRating' => null,
                 ],
             ]),
         ];