+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Added `\Wizaplace\SDK\Catalog\CompanySummary::isProfessional`
 - Added `\Wizaplace\SDK\Catalog\CompanySummary::getAverageRating`
 - Added `\Wizaplace\SDK\MailingList\MailingListService::isSubscribed`
+- Added `\Wizaplace\SDK\Catalog\Declination::getCompany`
 
 ## 0.8.0